+++ CHANGELOG.md
@@ -56,6 +56,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Update Lists ui
 - Update OpenedListPage : User can now click on an item
 - Update item page
+- Update user can now rename a wishlist and a recipe with a rename button
 
 ### Deprecated